modified:   app/integrations/integrationConfig.ts 	modified:   app/integrations/mealie/mealieShoppingLists.ts 	modified:   app/integrations/tandoor/tandoorShoppingLists.ts 	renamed:    server/utils/mealieShoppingLists.ts -> server/integrations/mealie/client.ts 	new file:   server/integrations/mealie/index.ts 	new file:   server/integrations/mealie/types.ts 	renamed:    server/utils/tandoorShoppingLists.ts -> server/integrations/tandoor/client.ts 	new file:   server/integrations/tandoor/index.ts 	new file:   server/integrations/tandoor/types.ts 	deleted:    server/tsconfig.json
