+++ server/integrations/mealie/index.ts
@@ -1,0 +1,5 @@
+// Export all types
+export * from './types'
+
+// Export the service class
+export { MealieService } from './client'
